fix(cgi): cast `RLIMIT_*` to `c_int` so the linux build typechecks
`libc::RLIMIT_AS` / `RLIMIT_CPU` / `RLIMIT_NPROC` are typed as
`__rlimit_resource_t` (u32) on linux glibc but as `c_int` (i32) on
macOS, so the previous code compiled on darwin and rejected the
ubuntu CI runner with three E0308 mismatches at the
`apply_rlimit` callsites in `crates/engine/src/fetch/cgi.rs`. Cast
each constant to `libc::c_int` at the callsite — a no-op on macOS
and a safe truncation on linux (the values are tiny enum
constants) — and leave `apply_rlimit`'s portable signature alone.
The function body already normalises the argument back to the
platform's `setrlimit` first-arg type via `as _`.

Local repro: `rustup target add x86_64-unknown-linux-gnu` and
`cargo check --target x86_64-unknown-linux-gnu -p vane-engine`.

+++ crates/engine/src/fetch/cgi.rs
@@ -779,12 +779,16 @@ fn pre_exec_drop_privs_and_limits(security: &CgiSecurity) -> io::Result<()> {
 	if unsafe { libc::setuid(security.uid as libc::uid_t) } != 0 {
 		return Err(io::Error::last_os_error());
 	}
+	// `libc::RLIMIT_*` is `c_int` (i32) on macOS and `__rlimit_resource_t`
+	// (u32) on Linux glibc. Cast at the callsite so `apply_rlimit`'s
+	// signature stays portable; the values themselves are small enum
+	// constants that fit in either type.
 	apply_rlimit(
-		libc::RLIMIT_AS,
+		libc::RLIMIT_AS as libc::c_int,
 		security.limits.memory_mb.map(|mb| mb.saturating_mul(1024 * 1024)),
 	)?;
-	apply_rlimit(libc::RLIMIT_CPU, security.limits.cpu_seconds)?;
-	apply_rlimit(libc::RLIMIT_NPROC, security.limits.max_processes)?;
+	apply_rlimit(libc::RLIMIT_CPU as libc::c_int, security.limits.cpu_seconds)?;
+	apply_rlimit(libc::RLIMIT_NPROC as libc::c_int, security.limits.max_processes)?;
 	Ok(())
 }
